requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
faiss-cpu
numpy
scikit-learn
rank-bm25
pydantic
python-dotenv

